requirement.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
pypdf
python-multipart
python-dotenv